fix: rename upload component and export it
